Fix a bug in periodic boundary condition (#45)
For the periodic boundary condition, the mesh has one grid less than the real computational domain. If the multi-level grid is used with a periodic boundary condition, the last layer grid of the finer mesh is outside of the coarse grid. This causes an interpolation issue and a mesh topology issue. This pull request fixes them by using the physical domain as the x[yz]max value. A machine tolerance is also added to help float number comparision.

* fix fine and corse mesh boundary

* Fix interpolation

* fix bugs

* add periodic pair check

* fix bugs

* Fix

* fix bug

* fix

* fix bugs

* fix bugs

* fix bugs

* fix bugs

* update couette inFlow

* optimise

* fix README.md

* adjust interpolation coeffcient

* fix bugs

* add machinetolence